+++ packages/alpha/src/__tests__/api/alertsApi.int.spec.ts
@@ -21,11 +21,38 @@ describe('alerts api', () => {
       },
     ]);
     expect(response.length).toBe(1);
+    expect(response[0].externalId).toBe(channelExternalId);
+  });
+
+  itif(client)('create channels with deduplication params', async () => {
+    const channelExternalIdWithDeduplication = channelExternalId + '_dedup';
+    const response = await client!.alerts.createChannels([
+      {
+        externalId: channelExternalIdWithDeduplication,
+        name: channelExternalIdWithDeduplication,
+        description: 'test with deduplication params',
+        alertRules: {
+          deduplication: {
+            activationInterval: '10m',
+            mergeInterval: '1m',
+          },
+        },
+      },
+    ]);
+    expect(response.length).toBe(1);
+    expect(response[0].externalId).toBe(channelExternalIdWithDeduplication);
+    expect(response[0].alertRules?.deduplication).toEqual({
+      activationInterval: '10m',
+      mergeInterval: '1m',
+    });
   });
 
   itif(client)('list channels', async () => {
-    const response = await client!.alerts.listChannels();
-    expect(response.items.length).toBeGreaterThan(0);
+    const response = await client!.alerts.listChannels({
+      filter: { externalIds: [channelExternalId] },
+    });
+    expect(response.items.length).toEqual(1);
+    expect(response.items[0].externalId).toBe(channelExternalId);
   });
 
   itif(client)('create alerts', async () => {
@@ -50,8 +77,12 @@ describe('alerts api', () => {
   });
 
   itif(client)('list alerts', async () => {
-    const response = await client!.alerts.list();
-    expect(response.items.length).toBeGreaterThan(0);
+    const response = await client!.alerts.list({
+      filter: {
+        channelExternalIds: [channelExternalId],
+      },
+    });
+    expect(response.items.length).toEqual(1);
   });
 
   itif(client)('create subscribers', async () => {
@@ -62,6 +93,17 @@ describe('alerts api', () => {
       },
     ]);
     expect(response.length).toBe(1);
+  });
+
+  itif(client)('list subscribers', async () => {
+    const response = await client!.alerts.listSubscribers({
+      filter: {
+        email,
+        externalIds: [email],
+      },
+    });
+    expect(response.items.length).toBe(1);
+    expect(response.items[0].externalId).toBe(email);
   });
 
   itif(client)('create subscriptions', async () => {
@@ -76,10 +118,47 @@ describe('alerts api', () => {
     expect(response.length).toBe(1);
   });
 
+  itif(client)('list subscriptions', async () => {
+    const response = await client!.alerts.listSubscriptions({
+      filter: {
+        channelExternalIds: [channelExternalId],
+        subscriberExternalIds: [email],
+        externalIds: [email],
+        metadata: { a: '1' },
+      },
+    });
+    expect(response.items.length).toBe(1);
+    expect(response.items[0].channelExternalId).toBe(channelExternalId);
+  });
+
   itif(client)('delete subscriptions', async () => {
     const response = await client!.alerts.deleteSubscription([
       {
         externalId: email,
+      },
+    ]);
+    expect(response).toEqual({});
+  });
+
+  itif(client)('delete subscriber', async () => {
+    const response = await client!.alerts.deleteSubscribers([
+      {
+        externalId: email,
+      },
+    ]);
+    expect(response).toEqual({});
+    const emptyRes = await client!.alerts.listSubscribers({
+      filter: {
+        email,
+      },
+    });
+    expect(emptyRes.items.length).toBe(0);
+  });
+
+  itif(client)('delete channel', async () => {
+    const response = await client!.alerts.deleteChannels([
+      {
+        externalId: channelExternalId,
       },
     ]);
     expect(response).toEqual({});