+++ packages/codegen/src/generator/generator.ts
@@ -1,6 +1,7 @@
 export interface TypeGeneratorResult {
   typeNames: string[];
   code: string;
+  astProcessedCode: string;
 }
 
 export interface TypeGenerator {