+++ packages/alpha/CHANGELOG.md
@@ -3,6 +3,117 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.1](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-alpha@0.4.0...@cognite/sdk-alpha@0.4.1) (2023-01-06)
+
+**Note:** Version bump only for package @cognite/sdk-alpha
+
+
+
+
+
+# [0.4.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-alpha@0.3.3...@cognite/sdk-alpha@0.4.0) (2022-12-15)
+
+
+### Features
+
+* **alpha:** add alert subscribers delete ([#923](https://github.com/cognitedata/cognite-sdk-js/issues/923)) ([0f378ca](https://github.com/cognitedata/cognite-sdk-js/commit/0f378caf1cf68acd9dcd2bfc289dd23aa40076fb))
+
+
+
+
+
+## [0.3.3](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-alpha@0.3.2...@cognite/sdk-alpha@0.3.3) (2022-12-13)
+
+
+### Bug Fixes
+
+* **alpha:** export cognite client properly ([#931](https://github.com/cognitedata/cognite-sdk-js/issues/931)) ([c4399eb](https://github.com/cognitedata/cognite-sdk-js/commit/c4399ebdfacbcc41ba14e35f90b3de8be1918462))
+
+
+
+
+
+## [0.3.2](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-alpha@0.3.1...@cognite/sdk-alpha@0.3.2) (2022-12-09)
+
+**Note:** Version bump only for package @cognite/sdk-alpha
+
+
+
+
+
+## [0.3.1](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-alpha@0.3.0...@cognite/sdk-alpha@0.3.1) (2022-12-02)
+
+
+### Bug Fixes
+
+* **alpha:** rename monitoringtask to monitoringtasks ([#924](https://github.com/cognitedata/cognite-sdk-js/issues/924)) ([d5b9390](https://github.com/cognitedata/cognite-sdk-js/commit/d5b9390dcbb288e060c5a0b4ff00df9a54a851a9))
+
+
+
+
+
+# [0.3.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-alpha@0.2.0...@cognite/sdk-alpha@0.3.0) (2022-12-01)
+
+
+### Features
+
+* **alpha:** monitoring tasks api ([#921](https://github.com/cognitedata/cognite-sdk-js/issues/921)) ([9161a69](https://github.com/cognitedata/cognite-sdk-js/commit/9161a69e2b0f4093099e0f36a79c0abefce6723c))
+
+
+
+
+
+# [0.2.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-alpha@0.1.17...@cognite/sdk-alpha@0.2.0) (2022-11-23)
+
+
+### Features
+
+* **alpha:** alerts list subscribers and subscriptions ([#920](https://github.com/cognitedata/cognite-sdk-js/issues/920)) ([ae6b04b](https://github.com/cognitedata/cognite-sdk-js/commit/ae6b04b115bf8d5c92cdb0108f1427de36a17c31))
+
+
+
+
+
+## [0.1.17](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-alpha@0.1.16...@cognite/sdk-alpha@0.1.17) (2022-11-21)
+
+**Note:** Version bump only for package @cognite/sdk-alpha
+
+
+
+
+
+## [0.1.16](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-alpha@0.1.15...@cognite/sdk-alpha@0.1.16) (2022-11-18)
+
+**Note:** Version bump only for package @cognite/sdk-alpha
+
+
+
+
+
+## [0.1.15](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-alpha@0.1.14...@cognite/sdk-alpha@0.1.15) (2022-11-17)
+
+**Note:** Version bump only for package @cognite/sdk-alpha
+
+
+
+
+
+## [0.1.14](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-alpha@0.1.13...@cognite/sdk-alpha@0.1.14) (2022-11-02)
+
+**Note:** Version bump only for package @cognite/sdk-alpha
+
+
+
+
+
+## [0.1.13](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-alpha@0.1.12...@cognite/sdk-alpha@0.1.13) (2022-10-28)
+
+**Note:** Version bump only for package @cognite/sdk-alpha
+
+
+
+
+
 ## [0.1.12](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-alpha@0.1.11...@cognite/sdk-alpha@0.1.12) (2022-10-04)
 
 **Note:** Version bump only for package @cognite/sdk-alpha